pyproject.toml
[project]
name = "ai-fintec-risk-intel"
version = "0.1.0"
description = "Portfolio-grade FinTech Risk Intelligence Platform (Fraud/Anomaly + Conduct/Compliance) with auditability and observability."
requires-python = ">=3.11,<3.13"
readme = "README.md"
license = { text = "Proprietary" }

dependencies = [
  "fastapi>=0.111.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.3.0",
  "orjson>=3.10.0",
  "httpx>=0.27.0",
  "python-json-logger>=2.0.7",

  "sqlalchemy>=2.0.30",
  "psycopg[binary]>=3.1.19",
  "alembic>=1.13.1",

  "boto3>=1.34.120",

  "opensearch-py>=2.6.0",

  "numpy>=1.26.4",
  "pandas>=2.2.2",
  "scikit-learn>=1.5.0",
  "joblib>=1.4.2",

  "shap>=0.46.0",

  "tenacity>=8.3.0",

  "openai>=1.40.0",

  "opentelemetry-api>=1.26.0",
  "opentelemetry-sdk>=1.26.0",
  "opentelemetry-exporter-otlp>=1.26.0",
  "opentelemetry-instrumentation-fastapi>=0.47b0",
  "opentelemetry-instrumentation-requests>=0.47b0",
  "opentelemetry-instrumentation-httpx>=0.47b0",

  "prometheus-client>=0.20.0",
  "langchain>=0.2.10",
"langchain-openai>=0.1.17",
"faiss-cpu>=1.8.0",
]

[project.optional-dependencies]
nlp = [
  "sentence-transformers>=3.0.1",
  "torch>=2.3.1",
]
dev = [
  "pytest>=8.2.0",
  "pytest-asyncio>=0.23.7",
  "ruff>=0.5.0",
  "mypy>=1.10.0",
  "types-requests>=2.32.0.20240602",
]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
no_implicit_optional = true
strict_optional = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests", "services", "libs"]

[build-system]
requires = ["setuptools>=70"]
build-backend = "setuptools.build_meta"